+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 ### Fixes
 
 - Fixed an issue causing encryption keys with no Key Ring value to populate with a `null` key ring
+- Fixed an issue where VPC address prefixes were not being updated when advanced subnets were created with Manual Address Management enabled
 - Fixed an issue preventing Terraform from being created when using an existing VPC
 - Fixed an issue causing Routing Tables to not have correct reference to an imported VPC